- use tm as a convenience library

+++ tests/tools/iterate.c
@@ -129,7 +129,7 @@ main(int argc, char **argv)
 			talloc_free(tmp);
 		}
 		if (when == cm_time_delay) {
-			printf("delay=%ld\n", delay);
+			printf("delay=%ld\n", (long) delay);
 		}
 		/* If we didn't find a match, stop here. */
 		if (*p == '\0') {
